hw 14 upload
Instructions .zip has all files, index.html and index.js are the ones I created for this assignment. Work semi-well, filters all work, but they do not build on eachother as I would like them to. I think a simple variable re-assignment would do the trick, just didnt get to it. Site is not paginated and takes forever to load.